+++ numeric-optimization/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to the
   the flag to build without L-BFGS.
 * Add some instances of standard type classes: `Eq OptimizationException`,
   `Show Result`, and `Show Statistics`.
+* Return correct statistics for L-BFGS and L-BFGS-B.
 
 ## 0.1.0.1 - 2023-06-03