+++ numeric-optimization/CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to the
 [Haskell Package Versioning Policy](https://pvp.haskell.org/).
 
-## 0.1.1.0 - Unreleased
+## 0.1.1.0 - 2023-06-21
 
 * Support L-BFGS-B algorithm (when `with-lbfgsb` is enabled)
 * Add some algorithm specific parameters